requirements.txt
Flask==3.0.3
gpxpy==1.6.2
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
matplotlib==3.8.4
